create color shades

For each color in col create as many shades as indicated by the value of reps at the corresponding index by grading towards 'white' - the shades are going scale of the way to white/black.

color.shades(
  reps,
  col = col.fun(length(reps)),
  scale = ifelse(direction == "bright", 0.7, 0.5),
  gscale = TRUE,
  direction = c("all", "bright", "dark"),
  col.fun = default.hcl,
  ...
)

Arguments

reps
vector of repetitions

col
color vector; must be a valid argument to col2rgb or a \(3\times N\) RGB color matrix (one column per color).

scale
how much of the range to black/white should be covered

gscale
global shade steps?

direction
which way to build the shades, see Description

...
passed to colorShadeRamp

Value

Color matrix with on entry per column

Details

If reps is scalar it is recycled to the length of col, otherwise it is filled with ones should it be shorter than col.

The shades can be constructed either towards white, black, or, symmetrically. This is determined by setting direction to 'bright', 'dark', or 'all', respectively. The latter is the default. When gscale is TRUE, all shade steps are the same, independent of the individual repetitions, otherwise each shade range covers the range given by scale.

Examples

color.shades(c('red', 'blue'), 3)

#> Error in max.rep%/%2: non-numeric argument to binary operator

show.shades(color.shades(c('red', 'green', 'blue'), 2:4))

#> Error in max.rep%/%2: non-numeric argument to binary operator